for all cohorts

+++ convert_to_tiff.py
@@ -21,29 +21,38 @@ log_config = os.path.join(os.path.split(__file__)[0], "resources", "log4j.proper
 # Parameters:
 level = 0  # image pyramid level to get size from
 
-CK_paths = "/data/Maren_P1/epithelium/CK/"
-geojson_path = "/home/maren/workspace/qupath-ck-seg/export_geojson_291122/"
-tmp_path = "/home/maren/workspace/qupath-ck-seg/tmp_results/"
-output_path = "/home/maren/workspace/qupath-ck-seg/pyramidal_tiff/"
+#CK_paths = "/data/Maren_P1/epithelium/CK/"
+#geojson_path = "/home/maren/workspace/qupath-ck-seg/export_geojson_291122/"
+#tmp_path = "/home/maren/workspace/qupath-ck-seg/tmp_results/"
+#output_path = "/home/maren/workspace/qupath-ck-seg/pyramidal_tiff/"
+
+# Need to run this script once for every folder with CK paths
+cohorts_path = "/data/Maren_P1/data/TMA/cohorts/"
+geojson_path = "/data/Maren_P1/data/annotations_converted/blue_channel_tumor_only/"
+tmp_path = "/data/Maren_P1/data/annotations_converted/blue_channel_tumor_only/blue_channel_temp/"
+output_path = "/data/Maren_P1/data/annotations_converted/blue_channel_tiff/"
 
 # delete folders
-shutil.rmtree(tmp_path)
-#shutil.rmtree(output_path)  # usually keep this, but had one in there already that I wanted to keep
+if os.path.exists(tmp_path):
+    shutil.rmtree(tmp_path)
+#shutil.rmtree(output_path)  # do not remove if images one wants to keep
 
 os.makedirs(tmp_path, exist_ok=True)
 os.makedirs(output_path, exist_ok=True)
 
 wsis = []
-for file in os.listdir(CK_paths):
-    if (".vsi" in file) and ("Overview" not in file):
-        wsis.append(CK_paths + file)
-
-# wsis = wsis[:-1]
+for cohort in os.listdir(cohorts_path):
+    curr_cohort_path = cohorts_path + cohort + "/"
+    if cohort in ["HPA"]:
+        continue
+    for file in os.listdir(curr_cohort_path):
+        full_path = curr_cohort_path + file
+        # get CK only
+        if (".vsi" in full_path) and ("Overview" not in full_path) and ("_CK_" in full_path):
+            wsis.append(full_path)
 
 for pathCK in tqdm(wsis, "WSI:"):
     print(pathCK)
-    if not pathCK.endswith("3.vsi"):
-        continue
 
     # get ID
     id_ = pathCK.split("/")[-1].split(".")[0]
@@ -66,11 +75,15 @@ for pathCK in tqdm(wsis, "WSI:"):
     # convert from geoJSON to tiled geoTIFF (TIFF)
     sp.check_call(
         ["gdal_rasterize", "-burn", "1", "-ts", str(width), str(height), "-te", "0", "0", str(width), str(height),
-         "-ot", "Byte", file1, file2, "-co", "COMPRESS=LZW"])
+         "-ot", "Byte", file1, file2, "-co", "COMPRESS=LZW"], stderr=sp.DEVNULL, stdout=sp.DEVNULL)
 
     # convert from TIFF to pyramidal TIFF
     sp.check_call(["vips", "tiffsave", file2, file3, "--bigtiff", "--tile", "--pyramid",
-                   "--compression=lzw"])  # do I need to change quality of compression? "--Q=" + str(Q2)?
+                   "--compression=lzw"], stderr=sp.DEVNULL, stdout=sp.DEVNULL)  # do I need to change quality of compression? "--Q=" + str(Q2)?
+
+# when finished, delete temporary dir
+if os.path.exists(tmp_path):
+    shutil.rmtree(tmp_path)
 
 # kill java VM
 jb.kill_vm()
